# Do not scan vendored copies of six.moves: Makes start faster