Skip modules that end with the name .six.moves since many libraries
include their own copy. This makes starting freezegun much faster.
Freezegun already skips the module named "six.moves" by appending
it to the freezer's "ignore" list. However, many libraries ship their
own copy, and import it under their own name. For example, in
pkg_resources:

    from pkg_resources.extern import six

A list of copies of six.py in a project I work on shows the following
libraries include a copy, so I suspect this is pretty common:

boto
pip
pkg_resources
urllib3