@evanj 2017-12-21
I used the following script to test this. To reproduce the problem, do:

1. `virtualenv bugvenv``
2. `bugvenv/bin/pip install freezegun`
3. `bugvenv/bin/python timings.py`

On my machine, without this change, I get the following timings:

```
BEFORE pkg_resources:
1000 iterations in 2.500854 s; 2.500854 ms / call
1000 iterations in 2.475269 s; 2.475269 ms / call
1000 iterations in 2.484204 s; 2.484204 ms / call

AFTER pkg_resources:
1000 iterations in 6.951676 s; 6.951676 ms / call
1000 iterations in 6.825846 s; 6.825846 ms / call
1000 iterations in 6.937931 s; 6.937931 ms / call
```

After this change, I get the following:

```
BEFORE pkg_resources:
1000 iterations in 2.555821 s; 2.555821 ms / call
1000 iterations in 2.595121 s; 2.595121 ms / call
1000 iterations in 2.513985 s; 2.513985 ms / call

AFTER pkg_resources:
1000 iterations in 4.956954 s; 4.956954 ms / call
1000 iterations in 5.005256 s; 5.005256 ms / call
1000 iterations in 5.007610 s; 5.007610 ms / call
```

In my real application's test suite, which has multiple copies of six, the difference is significantly more dramatic.